Move the LibraryMenu to its own file
